services/ai-engine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
yfinance>=0.2.40
fredapi==0.5.2
feedparser==6.0.11
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.0
openai==1.12.0
langchain==0.1.5
langchain-google-genai>=0.1.0 
google-generativeai>=0.8.0
websockets==12.0
websockets==12.0
python-telegram-bot>=20.0
